Move RenderPass to API, add Descriptions for Many Methods, and Adjust Some
