+++ common/src/main/java/org/vivecraft/client_vr/render/helpers/DebugRenderHelper.java
@@ -25,7 +25,7 @@ import org.vivecraft.client_vr.gameplay.trackers.DebugRenderTracker;
 import org.vivecraft.client_vr.gameplay.trackers.TelescopeTracker;
 import org.vivecraft.client_vr.provider.DeviceSource;
 import org.vivecraft.client_vr.provider.MCVR;
-import org.vivecraft.client_vr.render.RenderPass;
+import org.vivecraft.api.client.data.RenderPass;
 import org.vivecraft.common.utils.MathUtils;
 
 import java.lang.Math;
